move around next-components export
