+++ pages/_document.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Document, {Head, Main, NextScript} from 'next/document'
 import {ServerStyleSheet} from 'styled-components'
-import {getAssetPath, CommonStyles, CommonScripts} from '../next-components/utils'
+import {getAssetPath, CommonStyles, CommonScripts} from '../src/next-components/utils'
 
 export default class MyDocument extends Document {
   static getInitialProps({renderPage}) {
