+++ pages/_error.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {Heading} from '@primer/components'
-import {redirectTrailingSlash} from '../next-components'
+import {redirectTrailingSlash} from '../src/next-components'
 
 export default class Error extends React.Component {
   static getInitialProps(context) {
